Move BaseGapsSummaryReport into subclasses
