+++ custom/inddex/utils.py
@@ -111,28 +111,6 @@ class MultiTabularReport(DatespanMixin, CustomProjectReport, GenericTabularRepor
         return title, exported_rows
 
 
-class BaseGapsSummaryReport(MultiTabularReport):
-
-    @property
-    def fields(self):
-        return super().fields + [GapTypeFilter, RecallStatusFilter]
-
-    @property
-    def report_config(self):
-        report_config = super().report_config
-        report_config.update(gap_type=self.gap_type, recall_status=self.recall_status)
-
-        return report_config
-
-    @property
-    def gap_type(self):
-        return self.request.GET.get('gap_type') or ''
-
-    @property
-    def recall_status(self):
-        return self.request.GET.get('recall_status') or ''
-
-
 class BaseNutrientReport(MultiTabularReport):
 
     @property
